♻️ M1Z3D9 task: decompose finish execution
