+++ .agentplane/tasks/202605282057-M1Z3D9/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 M1Z3D9 task: Task finish execution decomposition [202605282057-M1Z3D9]
