+++ .agentplane/tasks/202605282057-M1Z3D9/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605282057-M1Z3D9/finish-execution-decomposition",
+  "created_at": "2026-05-28T20:57:51.056Z",
+  "diffstat_sha256": "sha256:7d3dbfc2324512e1db3825221ce9197f9b0a9ae70cf72eef08a9c3a6484155e0",
+  "last_verified_at": "2026-05-28T21:03:19.406Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605282057-M1Z3D9",
+  "updated_at": "2026-05-28T20:57:51.056Z",
+  "verify": {
+    "status": "pass"
+  }
+}
