+++ .agentplane/tasks/202605282057-M1Z3D9/pr/review.md
@@ -1,0 +1,40 @@
+# PR Review
+
+Created: 2026-05-28T20:57:51.056Z
+
+## Task
+
+- Task: `202605282057-M1Z3D9`
+- Title: Task finish execution decomposition
+- Status: DOING
+- Branch: `task/202605282057-M1Z3D9/finish-execution-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282057-M1Z3D9/README.md`
+
+## Verification
+
+- State: ok
+- Note: Task finish execution decomposition verified locally.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T20:57:51.056Z
+- Branch: task/202605282057-M1Z3D9/finish-execution-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../src/commands/task/finish-execute-close.ts      | 140 +++++++++
+ .../src/commands/task/finish-execute-commit.ts     | 105 +++++++
+ .../src/commands/task/finish-execute-load.ts       |  66 +++++
+ .../agentplane/src/commands/task/finish-execute.ts | 313 +--------------------
+ 4 files changed, 317 insertions(+), 307 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
