+++ packages/agentplane/src/commands/task/finish-execute.ts
@@ -1,7 +1,6 @@
-import { invalidValueMessage } from "../../cli/output.js";
 import { PolicyEngine } from "../../policy/engine.js";
 import { CliError } from "../../shared/errors.js";
-import { collectTaskIncidents, renderIncidentCollectionPlanOutcome } from "../incidents/shared.js";
+import { renderIncidentCollectionPlanOutcome } from "../incidents/shared.js";
 import { throwIfPolicyDecisionDenied } from "../shared/policy-deny.js";
 import type { CommandContext } from "../shared/task-backend.js";
 
@@ -10,34 +9,19 @@ import {
   assertQualityReviewBeforeFinish,
 } from "./finish-blueprint-evidence.js";
 import {
-  createTaskCloseCommit,
   existingCommitInfo,
   loadTaskForFinish,
   refreshAcrArtifactsForFinishedTasks,
   type LoadedFinishTask,
-  type ResolvedCommitInfo,
   writeFinishedTasks,
 } from "./finish-shared.js";
-import {
-  clearDirectWorkLockIfMatches,
-  materializeBranchPrCloseTail,
-  resolveBranchPrCloseTailState,
-} from "./finish-close.js";
+import { clearDirectWorkLockIfMatches, resolveBranchPrCloseTailState } from "./finish-close.js";
 import { appendFinishStructuredFinding } from "./finish-findings.js";
-import {
-  defaultCommitEmojiForStatus,
-  enforceStatusCommitPolicy,
-  prepareTaskTransitionComment,
-  readCommitInfo,
-  runTaskTransitionCommentCommit,
-} from "./shared.js";
+import { enforceStatusCommitPolicy } from "./shared.js";
 import type { FinishExecutionPlan, FinishOptions } from "./finish-types.js";
-import { exitCodeForError } from "../../cli/exit-codes.js";
-import {
-  gitMutationDiagnosticContext,
-  resolveGitIndexLockInfo,
-} from "../../shared/git-mutation.js";
-import { resolveIgnoredDirectCloseDirtyPaths } from "../guard/impl/close-dirt.js";
+import { collectIncidentsForLoadedTasks, loadFinishTasks } from "./finish-execute-load.js";
+import { resolveTaskCommitInfo } from "./finish-execute-commit.js";
+import { assertCloseCommitCanMutateTaskState, finalizeCloseTail } from "./finish-execute-close.js";
 
 export async function executeFinishPlan(opts: {
   ctx: CommandContext;
@@ -251,290 +235,5 @@ async function appendStructuredFindingIfNeeded(opts: {
     taskId: plan.primaryTaskId,
     author: options.author,
     finding: plan.finishFinding,
-  });
-}
-
-async function loadFinishTasks(opts: {
-  ctx: CommandContext;
-  options: FinishOptions;
-  plan: FinishExecutionPlan;
-}): Promise<{
-  loadedTasks: LoadedFinishTask[];
-  primaryStatusFrom: string | null;
-  primaryTag: string | null;
-}> {
-  const { ctx, options, plan } = opts;
-  let primaryStatusFrom: string | null = null;
-  let primaryTag: string | null = null;
-  const loadedTasks: LoadedFinishTask[] = [];
-
-  for (const taskId of options.taskIds) {
-    const loaded = await loadTaskForFinish({
-      ctx,
-      store: plan.store,
-      useStore: plan.useStore,
-      taskId,
-      taskCount: options.taskIds.length,
-      metaTaskId: plan.metaTaskId,
-      resultProvided: plan.resultProvided,
-      resultSummary: plan.resultSummary,
-      force: options.force,
-      capturePrimaryLifecycleMeta: taskId === plan.primaryTaskId,
-    });
-    loadedTasks.push(loaded.loaded);
-    if (taskId === plan.primaryTaskId) {
-      primaryStatusFrom = loaded.primaryStatusFrom;
-      primaryTag = loaded.primaryTag;
-    }
-  }
-
-  return { loadedTasks, primaryStatusFrom, primaryTag };
-}
-
-async function collectIncidentsForLoadedTasks(opts: {
-  ctx: CommandContext;
-  taskIds: string[];
-  loadedTasks: LoadedFinishTask[];
-  write: boolean;
-}): Promise<{
-  plans: Awaited<ReturnType<typeof collectTaskIncidents>>["plan"][];
-  registryPaths: string[][];
-}> {
-  const plans = [];
-  const registryPaths = [];
-  for (const taskId of opts.taskIds) {
-    const loadedTask = opts.loadedTasks.find((candidate) => candidate.taskId === taskId) ?? null;
-    const collected = await collectTaskIncidents({
-      ctx: opts.ctx,
-      taskId,
-      task: loadedTask?.task ?? null,
-      write: opts.write,
-    });
-    plans.push(collected.plan);
-    registryPaths.push(collected.registryPaths);
-  }
-  return { plans, registryPaths };
-}
-
-async function resolveTaskCommitInfo(opts: {
-  ctx: CommandContext;
-  options: FinishOptions;
-  plan: FinishExecutionPlan;
-  primaryStatusFrom: string | null;
-  primaryTag: string | null;
-}): Promise<ResolvedCommitInfo | null> {
-  const { ctx, options, plan, primaryStatusFrom, primaryTag } = opts;
-  let taskCommitInfo: ResolvedCommitInfo | null = options.commit
-    ? await readCommitInfo(ctx.resolvedProject.gitRoot, options.commit)
-    : null;
-  const preparedComment =
-    options.commitFromComment || plan.statusCommitRequested
-      ? prepareTaskTransitionComment({
-          body: options.body,
-          enabled: true,
-          config: ctx.config,
-        })
-      : null;
-
-  if (options.commitFromComment) {
-    if (typeof options.commitEmoji === "string" && options.commitEmoji.trim() !== "✅") {
-      throw new CliError({
-        exitCode: 2,
-        code: "E_USAGE",
-        message: invalidValueMessage(
-          "--commit-emoji",
-          options.commitEmoji,
-          "✅ (finish commits must use a checkmark)",
-        ),
-      });
-    }
-    taskCommitInfo = await runTaskTransitionCommentCommit({
-      ctx,
-      cwd: options.cwd,
-      rootOverride: options.rootOverride,
-      taskId: plan.primaryTaskId,
-      primaryTag: primaryTag ?? "meta",
-      author: options.author,
-      statusFrom: primaryStatusFrom ?? undefined,
-      statusTo: "DONE",
-      commentBody: options.body,
-      formattedComment: preparedComment?.formattedComment ?? null,
-      emoji: options.commitEmoji ?? defaultCommitEmojiForStatus("DONE"),
-      allow: options.commitAllow,
-      autoAllow: false,
-      allowTasks: options.commitAllowTasks,
-      requireClean: options.commitRequireClean,
-      quiet: options.quiet,
-      progressMessage: "creating commit from verification comment",
-      resolveExecutorAgent: true,
-    });
-  }
-
-  if (plan.statusCommitRequested) {
-    if (
-      typeof options.statusCommitEmoji === "string" &&
-      options.statusCommitEmoji.trim() !== "✅"
-    ) {
-      throw new CliError({
-        exitCode: 2,
-        code: "E_USAGE",
-        message: invalidValueMessage(
-          "--status-commit-emoji",
-          options.statusCommitEmoji,
-          "✅ (finish commits must use a checkmark)",
-        ),
-      });
-    }
-    await runTaskTransitionCommentCommit({
-      ctx,
-      cwd: options.cwd,
-      rootOverride: options.rootOverride,
-      taskId: plan.primaryTaskId,
-      primaryTag: primaryTag ?? "meta",
-      author: options.author,
-      statusFrom: primaryStatusFrom ?? undefined,
-      statusTo: "DONE",
-      commentBody: options.body,
-      formattedComment: preparedComment?.formattedComment ?? null,
-      emoji: options.statusCommitEmoji ?? defaultCommitEmojiForStatus("DONE"),
-      allow: options.statusCommitAllow,
-      autoAllow: false,
-      allowTasks: true,
-      requireClean: options.statusCommitRequireClean,
-      quiet: options.quiet,
-      progressMessage: "creating status commit",
-      resolveExecutorAgent: true,
-    });
-  }
-
-  return taskCommitInfo;
-}
-
-async function assertCloseCommitCanMutateTaskState(opts: {
-  ctx: CommandContext;
-  options: FinishOptions;
-  plan: FinishExecutionPlan;
-}): Promise<void> {
-  const { ctx, options, plan } = opts;
-  if (!plan.shouldCloseCommit) return;
-
-  const taskId = plan.primaryTaskId ?? options.taskIds[0];
-  const lockInfo = await resolveGitIndexLockInfo({ repoRoot: ctx.resolvedProject.gitRoot });
-  if (lockInfo) {
-    throw new CliError({
-      exitCode: exitCodeForError("E_GIT_LOCKED"),
-      code: "E_GIT_LOCKED",
-      message: [
-        "finish close commit cannot proceed while a Git index lock file exists.",
-        "Why: close commit creation would fail after the task is marked DONE.",
-        "Fix: inspect the active git process or run `agentplane doctor git-locks` before rerunning finish.",
-        `Lock: ${lockInfo.lockPath}`,
-      ].join("\n"),
-      context: gitMutationDiagnosticContext({
-        command: "finish",
-        cwd: options.cwd,
-        repoRoot: ctx.resolvedProject.gitRoot,
-        gitDir: lockInfo.gitDir,
-        workflowMode: ctx.config.workflow_mode,
-        mutationKind: "close_tail",
-        taskId,
-        indexLockPath: lockInfo.lockPath,
-        indexLockAgeMs: lockInfo.ageMs,
-        remediation: "Inspect or clear the git index lock before rerunning finish.",
-      }),
-    });
-  }
-
-  const staged = await ctx.git.statusStagedPaths();
-  if (staged.length > 0 && options.closeUnstageOthers !== true) {
-    throw new CliError({
-      exitCode: 5,
-      code: "E_GIT",
-      message: [
-        "finish --close-commit cannot proceed with a non-empty index.",
-        "Why: close commit creation would fail after the task is marked DONE.",
-        `Staged paths: ${staged.slice(0, 12).join(", ")}${staged.length > 12 ? ` (+${staged.length - 12} more)` : ""}`,
-        "Fix: commit or unstage unrelated paths before rerunning finish, or pass --close-unstage-others to let finish unstage non-task paths first.",
-        "Safe command: git status --short --untracked-files=no",
-      ].join("\n"),
-      context: {
-        command: "finish",
-        task_id: taskId,
-        reason_code: "git_close_commit_dirty_index",
-        staged_paths: staged,
-      },
-    });
-  }
-
-  const unstaged = await ctx.git.statusUnstagedTrackedPaths();
-  const ignored =
-    ctx.config.workflow_mode === "direct"
-      ? await resolveIgnoredDirectCloseDirtyPaths({ ctx, taskId })
-      : [];
-  const blockingUnstaged = unstaged.filter((relPath) => !ignored.includes(relPath));
-  if (blockingUnstaged.length === 0) return;
-
-  throw new CliError({
-    exitCode: 5,
-    code: "E_GIT",
-    message: [
-      "finish --close-commit requires a clean tracked working tree before mutating task state.",
-      `Why: deterministic ${ctx.config.workflow_mode} close commits stage only the finished task artifacts; other tracked changes would make commit creation fail after the task is marked DONE.`,
-      `Dirty tracked paths: ${blockingUnstaged.slice(0, 12).join(", ")}${blockingUnstaged.length > 12 ? ` (+${blockingUnstaged.length - 12} more)` : ""}`,
-      "Fix: commit, stash, or move unrelated lifecycle changes into a batch close branch before rerunning finish.",
-      "Safe command: git status --short --untracked-files=no",
-    ].join("\n"),
-  });
-}
-
-async function finalizeCloseTail(opts: {
-  ctx: CommandContext;
-  options: FinishOptions;
-  plan: FinishExecutionPlan;
-  primaryTaskId: string;
-  promotedIncidents: number;
-}): Promise<void> {
-  const { ctx, options, plan, primaryTaskId, promotedIncidents } = opts;
-  if (!plan.shouldCloseCommit || !primaryTaskId) return;
-
-  if (!options.quiet) {
-    process.stdout.write("task marked DONE; creating deterministic close commit\n");
-  }
-  const closeUnstageOthers = options.closeCommit === true && options.closeUnstageOthers === true;
-  if (ctx.config.workflow_mode === "branch_pr") {
-    const closeBranch = await materializeBranchPrCloseTail({
-      ctx,
-      cwd: options.cwd,
-      rootOverride: options.rootOverride,
-      taskId: primaryTaskId,
-      baseBranchOverride: options.baseBranchOverride,
-      quiet: options.quiet,
-      closeUnstageOthers,
-      allowPolicy: promotedIncidents > 0,
-      additionalTaskIds: plan.closeAdditionalTaskIds,
-    });
-    if (options.quiet) return;
-    if (closeBranch) {
-      process.stdout.write(
-        `branch_pr close tail ready on ${closeBranch}; push that branch and open it with task hosted-close-pr if hosted automation does not create the closure PR for you.\n`,
-      );
-      return;
-    }
-    process.stdout.write(
-      "branch_pr close tail already handled; skipping local task-close branch materialization.\n",
-    );
-    return;
-  }
-
-  await createTaskCloseCommit({
-    ctx,
-    cwd: options.cwd,
-    rootOverride: options.rootOverride,
-    taskId: primaryTaskId,
-    baseBranchOverride: options.baseBranchOverride,
-    quiet: options.quiet,
-    closeUnstageOthers,
-    allowPolicy: promotedIncidents > 0,
-    additionalTaskIds: plan.closeAdditionalTaskIds,
   });
 }
